Description of Kadane Algorithm
Given an integer array nums, find the contiguous subarray (containing at least one number) which has the largest sum and return its sum.

Kadane’s Algorithm is a popular and efficient method to solve the "maximum subarray sum" problem. The idea is to find the contiguous subarray (within a one-dimensional array of numbers) which has the largest sum.